backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.118.0
greenlet==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
pillow==10.4.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
reportlab==4.4.3
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
setuptools
